Another interface yet again bites the dust!
So the big motivation here is that in almost all cases these
layers are using the 802.15.4 MAC FCFs to set values and
read values, but the original stack has this awfully complex
layer of indirection just in case anyone would want to ever override
any of that behavior and couldn't just tweak the implementation.

I personally feel this is silly. Write it the way it should
be written for 99% of cases, and let the researcher hack the
other 1% of cases. I feel a single strong implementation of the
common approach is probably way easier to understand than an
uncomprehensible mess that can be infinitely extended.